tests authorised redis client and adds tempenv.json to gitignore in case renaming test goes wrong and leaves env file named as tempenv
